should be long

+++ src/main/java/com/ociweb/pronghorn/pipe/BlobWriter.java
@@ -55,7 +55,7 @@ public abstract class BlobWriter extends OutputStream implements DataOutput, App
 	 
 	 abstract public void writePackedShort(short value);
 	 
-	 abstract public void writeLong(int v);
+	 abstract public void writeLong(long v);
 	 
 	 abstract public void writeInt(int v);
